# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  IT: "IT",
  "Non-IT": "Non-IT",
} as const;

export interface Job {
  id: number;
  title: string;
  company: string;
  type: JobType;
  location: string;
  salaryMin?: number | null;
  salaryMax?: number | null;
  experience: string;
  description: string;
  isActive: boolean;
  createdAt: string;
}

export type CreateJobBodyType =
  (typeof CreateJobBodyType)[keyof typeof CreateJobBodyType];

export const CreateJobBodyType = {
  IT: "IT",
  "Non-IT": "Non-IT",
} as const;

export interface CreateJobBody {
  title: string;
  company: string;
  type: CreateJobBodyType;
  location: string;
  salaryMin?: number | null;
  salaryMax?: number | null;
  experience: string;
  description: string;
  isActive?: boolean;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  called: "called",
  selected: "selected",
  rejected: "rejected",
} as const;

export interface Application {
  id: number;
  jobId?: number | null;
  fullName: string;
  mobileNumber: string;
  email: string;
  location: string;
  experience: string;
  skills: string;
  resumeUrl?: string | null;
  status: ApplicationStatus;
  createdAt: string;
}

export type ApplicationWithJob = Application & {
  jobTitle?: string | null;
  jobCompany?: string | null;
};

export interface CreateApplicationBody {
  jobId?: number | null;
  fullName: string;
  mobileNumber: string;
  email: string;
  location: string;
  experience: string;
  skills: string;
  resume?: Blob | null;
}

export type UpdateApplicationStatusBodyStatus =
  (typeof UpdateApplicationStatusBodyStatus)[keyof typeof UpdateApplicationStatusBodyStatus];

export const UpdateApplicationStatusBodyStatus = {
  pending: "pending",
  called: "called",
  selected: "selected",
  rejected: "rejected",
} as const;

export interface UpdateApplicationStatusBody {
  status: UpdateApplicationStatusBodyStatus;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  username: string;
}

export interface JobStats {
  total: number;
  itJobs: number;
  nonItJobs: number;
  locations: string[];
}

export interface AdminStats {
  totalJobs: number;
  activeJobs: number;
  totalApplications: number;
  pendingApplications: number;
  selectedApplications: number;
  rejectedApplications: number;
  calledApplications: number;
}

export type ListJobsParams = {
  /**
   * Filter by job type
   */
  type?: ListJobsType;
  /**
   * Filter by location
   */
  location?: string;
  /**
   * Minimum salary filter
   */
  minSalary?: number;
  /**
   * Maximum salary filter
   */
  maxSalary?: number;
  /**
   * Experience level filter
   */
  experience?: string;
};

export type ListJobsType = (typeof ListJobsType)[keyof typeof ListJobsType];

export const ListJobsType = {
  IT: "IT",
  "Non-IT": "Non-IT",
  all: "all",
} as const;

export type AdminListApplicationsParams = {
  /**
   * Filter by job ID
   */
  jobId?: number;
  /**
   * Filter by status
   */
  status?: AdminListApplicationsStatus;
};

export type AdminListApplicationsStatus =
  (typeof AdminListApplicationsStatus)[keyof typeof AdminListApplicationsStatus];

export const AdminListApplicationsStatus = {
  pending: "pending",
  called: "called",
  selected: "selected",
  rejected: "rejected",
} as const;