# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all active jobs
 */
export const ListJobsQueryParams = zod.object({
  type: zod
    .enum(["IT", "Non-IT", "all"])
    .optional()
    .describe("Filter by job type"),
  location: zod.coerce.string().optional().describe("Filter by location"),
  minSalary: zod.coerce.number().optional().describe("Minimum salary filter"),
  maxSalary: zod.coerce.number().optional().describe("Maximum salary filter"),
  experience: zod.coerce
    .string()
    .optional()
    .describe("Experience level filter"),
});

export const ListJobsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  type: zod.enum(["IT", "Non-IT"]),
  location: zod.string(),
  salaryMin: zod.number().nullish(),
  salaryMax: zod.number().nullish(),
  experience: zod.string(),
  description: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});
export const ListJobsResponse = zod.array(ListJobsResponseItem);

/**
 * @summary Get a single job by ID
 */
export const GetJobParams = zod.object({
  id: zod.coerce.number(),
});

export const GetJobResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  type: zod.enum(["IT", "Non-IT"]),
  location: zod.string(),
  salaryMin: zod.number().nullish(),
  salaryMax: zod.number().nullish(),
  experience: zod.string(),
  description: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get job statistics summary
 */
export const GetJobStatsResponse = zod.object({
  total: zod.number(),
  itJobs: zod.number(),
  nonItJobs: zod.number(),
  locations: zod.array(zod.string()),
});

/**
 * @summary Submit a job application
 */
export const CreateApplicationBody = zod.object({
  jobId: zod.number().nullish(),
  fullName: zod.string(),
  mobileNumber: zod.string(),
  email: zod.string(),
  location: zod.string(),
  experience: zod.string(),
  skills: zod.string(),
  resume: zod.instanceof(File).nullish(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  username: zod.string(),
});

/**
 * @summary Admin - list all jobs
 */
export const AdminListJobsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  type: zod.enum(["IT", "Non-IT"]),
  location: zod.string(),
  salaryMin: zod.number().nullish(),
  salaryMax: zod.number().nullish(),
  experience: zod.string(),
  description: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});
export const AdminListJobsResponse = zod.array(AdminListJobsResponseItem);

/**
 * @summary Admin - create a job
 */
export const AdminCreateJobBody = zod.object({
  title: zod.string(),
  company: zod.string(),
  type: zod.enum(["IT", "Non-IT"]),
  location: zod.string(),
  salaryMin: zod.number().nullish(),
  salaryMax: zod.number().nullish(),
  experience: zod.string(),
  description: zod.string(),
  isActive: zod.boolean().optional(),
});

/**
 * @summary Admin - update a job
 */
export const AdminUpdateJobParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUpdateJobBody = zod.object({
  title: zod.string(),
  company: zod.string(),
  type: zod.enum(["IT", "Non-IT"]),
  location: zod.string(),
  salaryMin: zod.number().nullish(),
  salaryMax: zod.number().nullish(),
  experience: zod.string(),
  description: zod.string(),
  isActive: zod.boolean().optional(),
});

export const AdminUpdateJobResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  type: zod.enum(["IT", "Non-IT"]),
  location: zod.string(),
  salaryMin: zod.number().nullish(),
  salaryMax: zod.number().nullish(),
  experience: zod.string(),
  description: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Admin - delete a job
 */
export const AdminDeleteJobParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Admin - list all applications
 */
export const AdminListApplicationsQueryParams = zod.object({
  jobId: zod.coerce.number().optional().describe("Filter by job ID"),
  status: zod
    .enum(["pending", "called", "selected", "rejected"])
    .optional()
    .describe("Filter by status"),
});

export const AdminListApplicationsResponseItem = zod
  .object({
    id: zod.number(),
    jobId: zod.number().nullish(),
    fullName: zod.string(),
    mobileNumber: zod.string(),
    email: zod.string(),
    location: zod.string(),
    experience: zod.string(),
    skills: zod.string(),
    resumeUrl: zod.string().nullish(),
    status: zod.enum(["pending", "called", "selected", "rejected"]),
    createdAt: zod.string(),
  })
  .and(
    zod.object({
      jobTitle: zod.string().nullish(),
      jobCompany: zod.string().nullish(),
    }),
  );
export const AdminListApplicationsResponse = zod.array(
  AdminListApplicationsResponseItem,
);

/**
 * @summary Admin - update application status
 */
export const AdminUpdateApplicationStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUpdateApplicationStatusBody = zod.object({
  status: zod.enum(["pending", "called", "selected", "rejected"]),
});

export const AdminUpdateApplicationStatusResponse = zod.object({
  id: zod.number(),
  jobId: zod.number().nullish(),
  fullName: zod.string(),
  mobileNumber: zod.string(),
  email: zod.string(),
  location: zod.string(),
  experience: zod.string(),
  skills: zod.string(),
  resumeUrl: zod.string().nullish(),
  status: zod.enum(["pending", "called", "selected", "rejected"]),
  createdAt: zod.string(),
});

/**
 * @summary Admin dashboard stats
 */
export const AdminGetStatsResponse = zod.object({
  totalJobs: zod.number(),
  activeJobs: zod.number(),
  totalApplications: zod.number(),
  pendingApplications: zod.number(),
  selectedApplications: zod.number(),
  rejectedApplications: zod.number(),
  calledApplications: zod.number(),
});