# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalJobs: number;
  activeJobs: number;
  totalApplications: number;
  pendingApplications: number;
  selectedApplications: number;
  rejectedApplications: number;
  calledApplications: number;
}