# lib/api-zod/src/generated/types/applicationWithJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from "./application";

export type ApplicationWithJob = Application & {
  jobTitle?: string | null;
  jobCompany?: string | null;
};