# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationBody {
  jobId?: number | null;
  fullName: string;
  mobileNumber: string;
  email: string;
  location: string;
  experience: string;
  skills: string;
  resume?: Blob | null;
}