# lib/api-zod/src/generated/types/createJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJobBodyType } from "./createJobBodyType";

export interface CreateJobBody {
  title: string;
  company: string;
  type: CreateJobBodyType;
  location: string;
  salaryMin?: number | null;
  salaryMax?: number | null;
  experience: string;
  description: string;
  isActive?: boolean;
}