# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { JobType } from "./jobType";

export interface Job {
  id: number;
  title: string;
  company: string;
  type: JobType;
  location: string;
  salaryMin?: number | null;
  salaryMax?: number | null;
  experience: string;
  description: string;
  isActive: boolean;
  createdAt: string;
}