# lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simple Samaj Job Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface JobStats {
  total: number;
  itJobs: number;
  nonItJobs: number;
  locations: string[];
}